requirements.txt
transformers==4.30.2
sentence-transformers==2.2.2
scikit-learn==1.3.0
torch==2.0.1
PyGithub==1.58.1